requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
mypy>=0.900
pylint>=2.12.0

# Type stubs for better IDE support
types-requests>=2.28.0
pandas-stubs>=1.2.0
types-PyYAML>=6.0.0